[View] Moving Renderer classes to a new Renderer namespace for better organization.
